refactor(domain): move colocation logic to use case
