+++ ATVersionControl/Classes/Domain/Model/ColocationEndpoint.swift
@@ -4,12 +4,7 @@
 //
 //  Created by Amir on 7/13/26.
 //
-public struct ColocationEndpoint: Decodable, Sendable {
+public struct ColocationEndpoint: Sendable {
     public var name = ""
     public var baseURL = ""
-
-    enum CodingKeys: String, CodingKey {
-        case name = "Name"
-        case baseURL = "BaseUrl"
-    }
 }
