+++ ATVersionControl/Classes/Domain/Model/VersionInfo.swift
@@ -6,9 +6,7 @@
 //  Copyright © 2019 ayantech.ir. All rights reserved.
 //
 
-import Foundation
-
-public struct VersionInfo: Decodable, Sendable {
+public struct VersionInfo: Sendable {
     public var acceptButtonText = ""
     public var body = ""
     public var changeLogs = [String]()
@@ -17,15 +15,4 @@ public struct VersionInfo: Decodable, Sendable {
     public var rejectButtonText = ""
     public var textToShare = ""
     public var title = ""
-
-    enum CodingKeys: String, CodingKey {
-        case acceptButtonText = "AcceptButtonText"
-        case body = "Body"
-        case changeLogs = "ChangeLogs"
-        case link = "Link"
-        case linkType = "LinkType"
-        case rejectButtonText = "RejectButtonText"
-        case textToShare = "TextToShare"
-        case title = "Title"
-    }
 }
