+++ ATVersionControl/Classes/Domain/Repository/ColocationLaneStoring.swift
@@ -1,0 +1,9 @@
+//
+//  ColocationLaneStoring.swift
+//  ATVersionControl
+//
+
+protocol ColocationLaneStoring {
+    func load() -> ColocationLane?
+    func save(_ lane: ColocationLane)
+}
